import djgpp 2.01

+++ src/libc/ansi/stdlib/strtod.c
@@ -1,3 +1,4 @@
+/* Copyright (C) 1996 DJ Delorie, see COPYING.DJ for details */
 /* Copyright (C) 1994 DJ Delorie, see COPYING.DJ for details */
 #include <math.h>
 #include <stdlib.h>
@@ -6,16 +7,16 @@
 double
 strtod(const char *s, char **sret)
 {
-  double r;			/* result */
+  long double r;		/* result */
   int e;			/* exponent */
-  double d;			/* scale */
+  long double d;		/* scale */
   int sign;			/* +- 1.0 */
   int esign;
   int i;
   int flags=0;
 
   r = 0.0;
-  sign = 1.0;
+  sign = 1;
   e = 0;
   esign = 1;
 
@@ -40,14 +41,14 @@ strtod(const char *s, char **sret)
 
   if (*s == '.')
   {
-    d = 0.1;
+    d = 0.1L;
     s++;
     while ((*s >= '0') && (*s <= '9'))
     {
       flags |= 2;
       r += d * (*s - '0');
       s++;
-      d *= 0.1;
+      d *= 0.1L;
     }
   }
 
@@ -77,7 +78,7 @@ strtod(const char *s, char **sret)
 
     while ((*s >= '0') && (*s <= '9'))
     {
-      e *= 10.0;
+      e *= 10;
       e += *s - '0';
       s++;
     }
@@ -85,7 +86,7 @@ strtod(const char *s, char **sret)
 
   if (esign < 0)
     for (i = 1; i <= e; i++)
-      r *= 0.1;
+      r *= 0.1L;
   else
     for (i = 1; i <= e; i++)
       r *= 10.0;
